feat: add ability and new permissions for every endpoint

+++ apps/api/src/routes/management.ts
@@ -1,3 +1,4 @@
+import { SYSTEM_ROLES } from "@/libs/ability"
 import { adminMiddleware } from "@/middlewares/admin-middleware"
 import {
 	getUserAppAccessStatus,
@@ -21,6 +22,7 @@ import {
 	assignRoleToUser,
 	createRole,
 	deleteRole,
+	getRoleDetails,
 	getUserAppRoles,
 	getUserRolesForApp,
 	listRolesForApp,
@@ -52,30 +54,55 @@ export const manageRouter = new Hono()
 	})
 
 	.get("/user", async (c) => {
-		const ownerId = c.get("ownerId")
-		const userDetails = await getUserDetails({ userId: ownerId })
-		return c.json({ ...userDetails, isOwner: true })
+		const rules = c.get("abilityRules")
+		const adminId = c.get("adminUserId")
+		const userDetails = await getUserDetails({ userId: adminId })
+		return c.json({ ...userDetails, rules })
 	})
 
 	.get("/users/:id", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("view_user", "User")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização suficiente para essa ação"
+			})
+		}
 		const requestedUserId = c.req.param("id")
 		const userDetails = await getUserDetails({ userId: requestedUserId })
 		return c.json(userDetails)
 	})
 
 	.get("/users/:id/status", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("view_user", "User")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização suficiente para essa ação"
+			})
+		}
 		const requestedUserId = c.req.param("id")
 		const status = await getUserGlobalStatus({ userId: requestedUserId })
 		return c.json({ userId: requestedUserId, status })
 	})
 
 	.get("/users/:id/applications-roles", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("view_user", "User")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização suficiente para essa ação"
+			})
+		}
 		const requestedUserId = c.req.param("id")
 		const appRolesList = await getUserAppRoles({ userId: requestedUserId })
 		return c.json(appRolesList)
 	})
 
 	.get("/users/:id/applications-access", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("view_user", "User")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização suficiente para essa ação"
+			})
+		}
 		const requestedUserId = c.req.param("id")
 		const appAccessList = await getUserAppAccessStatuses({ userId: requestedUserId })
 		return c.json(appAccessList)
@@ -94,6 +121,12 @@ export const manageRouter = new Hono()
 			})
 		),
 		async (c) => {
+			const ability = c.get("ability")
+			if (ability.cannot("view_user", "User")) {
+				throw new HTTPException(403, {
+					message: "Você não tem autorização suficiente para essa ação"
+				})
+			}
 			const { appId } = c.req.valid("query")
 			const requestedUserId = c.req.param("id")
 			const roles = await getUserRolesForApp({ appId, userId: requestedUserId })
@@ -114,6 +147,12 @@ export const manageRouter = new Hono()
 			})
 		),
 		async (c) => {
+			const ability = c.get("ability")
+			if (ability.cannot("view_application", "Application")) {
+				throw new HTTPException(403, {
+					message: "Você não tem autorização suficiente para essa ação"
+				})
+			}
 			const { appId } = c.req.valid("query")
 			const requestedUserId = c.req.param("id")
 			const status = await getUserAppAccessStatus({ appId, userId: requestedUserId })
@@ -122,6 +161,12 @@ export const manageRouter = new Hono()
 	)
 
 	.get("/applications/:appId/redirect-uris", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("view_application", "Application")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização suficiente para essa ação"
+			})
+		}
 		const appId = c.req.param("appId")
 		const uris = await listRedirectUrisForApp({ appId })
 		return c.json(uris)
@@ -131,6 +176,12 @@ export const manageRouter = new Hono()
 		"/applications/:appId/redirect-uris",
 		zValidator("json", z.object({ uri: z.string().url() })),
 		async (c) => {
+			const ability = c.get("ability")
+			if (ability.cannot("create_redirect_uri", "Application")) {
+				throw new HTTPException(403, {
+					message: "Você não tem autorização para criar uma URI de Redirecionamento"
+				})
+			}
 			const appId = c.req.param("appId")
 			const { uri } = c.req.valid("json")
 			await addRedirectUri({ appId, uri })
@@ -142,6 +193,12 @@ export const manageRouter = new Hono()
 		"/applications/:appId/redirect-uris",
 		zValidator("json", z.object({ uriId: z.string().uuid() })),
 		async (c) => {
+			const ability = c.get("ability")
+			if (ability.cannot("delete_redirect_uri", "Application")) {
+				throw new HTTPException(403, {
+					message: "Você não tem autorização para excluir uma URI de Redirecionamento"
+				})
+			}
 			const appId = c.req.param("appId")
 			const { uriId } = c.req.valid("json")
 			await deleteRedirectUri({ appId, uriId })
@@ -150,6 +207,12 @@ export const manageRouter = new Hono()
 	)
 
 	.get("/users", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("view_user", "User")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização suficiente para essa ação"
+			})
+		}
 		const usersList = await listUsers()
 		return c.json(usersList)
 	})
@@ -158,6 +221,12 @@ export const manageRouter = new Hono()
 		"/users/:id/status",
 		zValidator("json", z.object({ status: z.enum(["active", "inactive"]) })),
 		async (c) => {
+			const ability = c.get("ability")
+			if (ability.cannot("edit_user_global_status", "User")) {
+				throw new HTTPException(403, {
+					message: "Você não tem autorização para modificar o status do usuário"
+				})
+			}
 			const { status } = c.req.valid("json")
 			const requestedUserId = c.req.param("id")
 			await setUserGlobalStatus({ userId: requestedUserId, status })
@@ -166,11 +235,23 @@ export const manageRouter = new Hono()
 	)
 
 	.get("/applications", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("view_application", "Application")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização suficiente para essa ação"
+			})
+		}
 		const apps = await listApplicationsWithCounts()
 		return c.json(apps)
 	})
 
 	.get("/applications/:appId", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("view_application", "Application")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização suficiente para essa ação"
+			})
+		}
 		const appId = c.req.param("appId")
 		const appDetails = await getApplicationDetails({ appId })
 		return c.json(appDetails)
@@ -190,6 +271,12 @@ export const manageRouter = new Hono()
 			})
 		),
 		async (c) => {
+			const ability = c.get("ability")
+			if (ability.cannot("create_application", "Application")) {
+				throw new HTTPException(403, {
+					message: "Você não tem autorização para criar uma aplicação"
+				})
+			}
 			const data = c.req.valid("json")
 			await createApplication(data)
 			return c.body(null, 201)
@@ -197,6 +284,12 @@ export const manageRouter = new Hono()
 	)
 
 	.delete("/:appId", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("delete_application", "Application")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização para excluir uma aplicação"
+			})
+		}
 		const appId = c.req.param("appId")
 		const coreAppId = await getConfigValue(CONFIG_KEYS.PRIMARY_MANAGEMENT_APP_ID)
 		if (appId === coreAppId) {
@@ -207,6 +300,12 @@ export const manageRouter = new Hono()
 	})
 
 	.get("/applications/:appId/roles", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("view_application", "Application")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização suficiente para essa ação"
+			})
+		}
 		const appId = c.req.param("appId")
 		const rolesList = await listRolesForApp({ appId })
 		return c.json(rolesList)
@@ -216,9 +315,21 @@ export const manageRouter = new Hono()
 		"/applications/:appId/roles",
 		zValidator("json", z.object({ roleName: z.string().min(3).max(50) })),
 		async (c) => {
+			const ability = c.get("ability")
 			const appId = c.req.param("appId")
 			const { roleName } = c.req.valid("json")
-			await createRole({ appId, roleName: roleName })
+			if (ability.cannot("create_role", "Application")) {
+				throw new HTTPException(403, {
+					message: "Você não tem autorização para criar um cargo"
+				})
+			}
+			const managementAppId = await getConfigValue(CONFIG_KEYS.PRIMARY_MANAGEMENT_APP_ID)
+			if (appId === managementAppId && Object.values(SYSTEM_ROLES).includes(roleName)) {
+				throw new HTTPException(409, {
+					message: `O nome de cargo '${roleName}' é reservado para o sistema.`
+				})
+			}
+			await createRole({ appId, roleName })
 			return c.body(null, 201)
 		}
 	)
@@ -227,16 +338,48 @@ export const manageRouter = new Hono()
 		"/applications/:appId/roles/:roleId",
 		zValidator("json", z.object({ roleName: z.string().min(3).max(50) })),
 		async (c) => {
+			const ability = c.get("ability")
 			const appId = c.req.param("appId")
 			const roleId = c.req.param("roleId")
 			const { roleName } = c.req.valid("json")
-			await updateRole({ appId, roleId, roleName: roleName })
+			if (ability.cannot("edit_role", "Application")) {
+				throw new HTTPException(403, {
+					message: "Você não tem autorização para modificar um cargo"
+				})
+			}
+			const managementAppId = await getConfigValue(CONFIG_KEYS.PRIMARY_MANAGEMENT_APP_ID)
+			const roleDetails = await getRoleDetails({ roleId })
+			if (
+				roleDetails.appId === managementAppId &&
+				Object.values(SYSTEM_ROLES).includes(roleDetails.roleName)
+			) {
+				throw new HTTPException(403, {
+					message: "Não é possível modificar diretamente o nome de cargos críticos do sistema."
+				})
+			}
+			await updateRole({ appId, roleId, roleName })
 			return c.body(null, 204)
 		}
 	)
 
 	.delete("/roles/:roleId", async (c) => {
+		const ability = c.get("ability")
 		const roleId = c.req.param("roleId")
+		if (ability.cannot("delete_role", "Application")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização para apagar um cargo"
+			})
+		}
+		const roleToBeDeleted = await getRoleDetails({ roleId })
+		const managementAppId = await getConfigValue(CONFIG_KEYS.PRIMARY_MANAGEMENT_APP_ID)
+		if (
+			roleToBeDeleted.appId === managementAppId &&
+			Object.values(SYSTEM_ROLES).includes(roleToBeDeleted.roleName)
+		) {
+			throw new HTTPException(403, {
+				message: "Não é possível apagar cargos críticos do sistema."
+			})
+		}
 		await deleteRole({ roleId })
 		return c.body(null, 204)
 	})
@@ -245,6 +388,12 @@ export const manageRouter = new Hono()
 		"/users/:userId/roles",
 		zValidator("json", z.object({ roleId: z.string().uuid() })),
 		async (c) => {
+			const ability = c.get("ability")
+			if (ability.cannot("assign_role_to_user", "User")) {
+				throw new HTTPException(403, {
+					message: "Você não tem autorização para dar cargo ao usuário"
+				})
+			}
 			const userId = c.req.param("userId")
 			const { roleId } = c.req.valid("json")
 			await assignRoleToUser({ userId, roleId })
@@ -253,6 +402,12 @@ export const manageRouter = new Hono()
 	)
 
 	.delete("/users/:userId/roles/:roleId", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("revoke_role_from_user", "User")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização para revogar cargo do usuário"
+			})
+		}
 		const userId = c.req.param("userId")
 		const roleId = c.req.param("roleId")
 		await revokeRoleFromUser({ userId, roleId })
@@ -263,6 +418,12 @@ export const manageRouter = new Hono()
 		"/users/:userId/access/:appId",
 		zValidator("json", z.object({ status: z.enum(["enabled", "disabled"]) })),
 		async (c) => {
+			const ability = c.get("ability")
+			if (ability.cannot("edit_user_application_access", "User")) {
+				throw new HTTPException(403, {
+					message: "Você não tem autorização para modificar o status do usuário"
+				})
+			}
 			const appId = c.req.param("appId")
 			const userId = c.req.param("userId")
 			const { status } = c.req.valid("json")
@@ -272,18 +433,36 @@ export const manageRouter = new Hono()
 	)
 
 	.get("/:appId/api-keys", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("view_application", "Application")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização suficiente para essa ação"
+			})
+		}
 		const appId = c.req.param("appId")
 		const keys = await listApiKeysForApp({ appId })
 		return c.json(keys)
 	})
 
 	.post("/:appId/api-keys", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("create_api_key", "Application")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização para criar API Keys"
+			})
+		}
 		const appId = c.req.param("appId")
 		const apiKey = await generateApiKey({ appId })
 		return c.json(apiKey, 201)
 	})
 
 	.delete("/api-keys/:keyId", async (c) => {
+		const ability = c.get("ability")
+		if (ability.cannot("delete_api_key", "Application")) {
+			throw new HTTPException(403, {
+				message: "Você não tem autorização para deletar API Keys"
+			})
+		}
 		const keyId = c.req.param("keyId")
 		await revokeApiKey({ keyId })
 		return c.body(null, 204)
